backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "AI Data Analyst"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "duckdb>=1.3.0",
    "ipykernel>=6.29.5",
    "openai>=1.82.1",
    "pandas>=2.2.3",
    "python-dotenv>=1.1.0",
]
